# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "token传递方式包括[get/post]token、[header]Authorization\n/api/* 公共访问\n/adm/* 必须传入 token",
        "title": "Blog’s 接口文档",
        "contact": {},
        "version": "1.0"
    },
    "host": "127.0.0.1:8085",
    "basePath": "/",
    "paths": {
        "/adm/auth/edit": {
            "post": {
                "tags": [
                    "auth-登陆认证"
                ],
                "summary": "修改个人信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "名称",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "号码",
                        "name": "phone",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "邮箱",
                        "name": "email",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回数据",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Reply"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/adm/auth/get": {
            "get": {
                "tags": [
                    "auth-登陆认证"
                ],
                "summary": "获取登录信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "凭证",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回数据",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Reply"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/model.User"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/adm/auth/passwd": {
            "post": {
                "tags": [
                    "auth-登陆认证"
                ],
                "summary": "修改自己的密码",
                "parameters": [
                    {
                        "type": "string",
                        "description": "旧密码",
                        "name": "opasswd",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "新密码",
                        "name": "npasswd",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回数据",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Reply"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/adm/upload/file": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "ctrl-系统相关"
                ],
                "summary": "上传文件",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/adm/upload/image": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "ctrl-系统相关"
                ],
                "summary": "上传图片并裁剪",
                "parameters": [
                    {
                        "type": "string",
                        "description": "token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/auth/login": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "auth-登陆认证"
                ],
                "summary": "登陆",
                "parameters": [
                    {
                        "type": "string",
                        "default": "zxysilent",
                        "description": "账号",
                        "name": "num",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "zxyslt",
                        "description": "密码",
                        "name": "pass",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "返回数据",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Reply"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "auth-登陆认证"
                ],
                "summary": "注销登录",
                "responses": {}
            }
        },
        "/api/auth/vcode": {
            "get": {
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "auth-登陆认证"
                ],
                "summary": "验证码",
                "responses": {
                    "200": {
                        "description": "返回数据",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/model.Reply"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "string"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "model.Reply": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "msg": {
                    "type": "string",
                    "example": "提示信息"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "ctime": {
                    "description": "创建时间",
                    "type": "string"
                },
                "ecount": {
                    "description": "错误次数",
                    "type": "integer"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "id": {
                    "description": "主键",
                    "type": "integer"
                },
                "ltime": {
                    "description": "上次登录时间",
                    "type": "string"
                },
                "name": {
                    "description": "姓名",
                    "type": "string"
                },
                "num": {
                    "description": "账号",
                    "type": "string"
                },
                "openid_qq": {
                    "description": "qq",
                    "type": "string"
                },
                "passwd": {
                    "description": "密码",
                    "type": "string"
                },
                "phone": {
                    "description": "电话号码",
                    "type": "string"
                }
            }
        }
    }
}